// src/components/TableOld/TableOld.types.ts
export interface DataTableViewStateOld {
  columns: TableColumnsStateOld;
  sort: SortRuleOld[];
  filters: TableFilterOld[];
  query: string;
}